fix: Drop maxConcurrentRequests down to 2
Even though 10 seemed to be working fine for just updating the scores
for the most recent month for all languages, for whatever reason when
adding new languages and then running `update-data` the Stack Overflow
abuse prevention kept kicking in even with `maxConcurrentRequests` as
low as 3, possibly because it could end up calling the API a lot more.

For the time being it's probably going to be simplest just to leave it
at a low value since `update-data` will only typically run monthly and
there's no real impact if it takes a little bit longer.