feat: Remove functionality to get scores in parallel
Unfortunately, any value of `maxConcurrentRequests` over 1 triggers the Stack Overflow abuse
prevention when populating scores for multiple dates. We could keep it for populating scores
for one date, but we only do that during the monthly update and the value of saving a little
bit of time on a scheduled task that runs without any human intervention is lower than making
the code a bit simpler.

Closes #15