feat: Get total points from API below a certain threshold
This allows us to account for languages that are losing points (ActionScript, CoffeeScript, Classic ASP, etc). Keep the error checking logic because it actually seems to be useful; it caught an error with the Prolog data while I was determining the threshold to use for the new logic, which I had originally set to settings.minimumScore.