feat: Don't report missing languages until they approach minimum score
Adding new languages is a lot of work, and there's little value to keep adding every language if many of them aren't even going to accumulate enough points to show up in the chart. This should save quite a bit of work. Also remove Jai so we can have a better chance of remembering to add it later if it ever takes off.