test: Fix validateLanguages test
Previously it was timing out because we're no longer adding every language to the metadata, so validateLanguages takes longer to run because for every language not in the metadata that is in GitHub it will fetch the score to decide whether or not to return a message. Instead we override the GitHub languages to match our test metadata so that validateLanguages does not have to calculate any scores and also shouldn't fail the test if a language is missing.